Refactoring of _pyco_tree: * Improved compile time using type erasure. * Added the save_kd_tree and load_kd_tree functions. * Replaced several std::runtime_error exceptions by std::invalid_argument. Updated the minimum Python version to 3.10. Avoid global namespace redeclaration of index error with GNU extensions.

+++ examples/kd_tree/kd_tree_search.cpp
@@ -93,7 +93,7 @@ void search_s1() {
   std::cout << "Closest angles (index, distance, value): " << std::endl;
   for (auto const& nn : knn) {
     std::cout << "  " << nn.index << ", " << nn.distance << ", "
-              << tree.space()[nn.index] << std::endl;
+              << tree.space()[static_cast<std::size_t>(nn.index)] << std::endl;
   }
 }
